[stateless_validaiton] Refactor chunk_validator code
