+++ chain/client/src/lib.rs
@@ -23,13 +23,13 @@ pub use near_client_primitives::debug::DebugStatus;
 pub mod adapter;
 pub mod adversarial;
 mod chunk_inclusion_tracker;
-mod chunk_validation;
 mod client;
 mod client_actor;
 mod config_updater;
 pub mod debug;
 mod info;
 mod metrics;
+mod stateless_validation;
 pub mod sync;
 mod sync_jobs_actor;
 pub mod test_utils;
