+++ core/primitives/src/lib.rs
@@ -10,7 +10,6 @@ pub mod block;
 pub mod block_body;
 pub mod block_header;
 pub mod challenge;
-pub mod chunk_validation;
 pub mod epoch_manager;
 pub mod epoch_sync;
 pub mod errors;
@@ -28,6 +27,7 @@ pub mod state;
 pub mod state_part;
 pub mod state_record;
 pub mod state_sync;
+pub mod stateless_validation;
 pub mod static_clock;
 pub mod telemetry;
 pub mod test_utils;
